+++ src/onegov.org/onegov/org/forms/__init__.py
@@ -1,29 +1,24 @@
-from onegov.org.forms.allocation import (
-    DaypassAllocationForm,
-    DaypassAllocationEditForm,
-    RoomAllocationForm,
-    RoomAllocationEditForm
-)
+from onegov.org.forms.allocation import DaypassAllocationEditForm
+from onegov.org.forms.allocation import DaypassAllocationForm
+from onegov.org.forms.allocation import RoomAllocationEditForm
+from onegov.org.forms.allocation import RoomAllocationForm
 from onegov.org.forms.event import EventForm
 from onegov.org.forms.form_definition import CustomDefinitionForm
+from onegov.org.forms.generic import ExportForm, DateRangeForm
 from onegov.org.forms.imageset import ImageSetForm
 from onegov.org.forms.newsletter import NewsletterForm, NewsletterSendForm
 from onegov.org.forms.page import LinkForm, PageForm
 from onegov.org.forms.person import PersonForm
-from onegov.org.forms.reset_password import (
-    RequestPasswordResetForm,
-    PasswordResetForm
-)
 from onegov.org.forms.reservation import ReservationForm
-from onegov.org.forms.resource import (
-    ResourceForm,
-    ResourceCleanupForm,
-    ResourceExportForm
-)
+from onegov.org.forms.reset_password import PasswordResetForm
+from onegov.org.forms.reset_password import RequestPasswordResetForm
+from onegov.org.forms.resource import ResourceCleanupForm
+from onegov.org.forms.resource import ResourceExportForm
+from onegov.org.forms.resource import ResourceForm
 from onegov.org.forms.settings import SettingsForm
 from onegov.org.forms.signup import SignupForm
-from onegov.org.forms.userprofile import UserProfileForm
 from onegov.org.forms.user import ManageUserForm, NewUserForm
+from onegov.org.forms.userprofile import UserProfileForm
 
 
 __all__ = [
@@ -31,6 +26,8 @@ __all__ = [
     'DaypassAllocationEditForm',
     'DaypassAllocationForm',
     'EventForm',
+    'DateRangeForm',
+    'ExportForm',
     'ImageSetForm',
     'LinkForm',
     'ManageUserForm',