+++ src/onegov.org/onegov/org/views/occurrence.py
@@ -1,19 +1,18 @@
 """ The onegov org collection of images uploaded to the site. """
 
-from csv import writer
+from collections import OrderedDict
 from datetime import date
-from io import StringIO
 from morepath.request import Response
-from onegov.core.security import Public
+from onegov.core.security import Public, Private
 from onegov.core.utils import linkify
-from onegov.event import Event, Occurrence, OccurrenceCollection
+from onegov.event import Occurrence, OccurrenceCollection
 from onegov.org import _, OrgApp
 from onegov.org.elements import Link
+from onegov.org.forms import ExportForm
 from onegov.org.layout import OccurrenceLayout, OccurrencesLayout
 from onegov.ticket import TicketCollection
 from sedate import as_datetime, replace_timezone
-from sqlalchemy import func
-from sqlalchemy.dialects.postgresql import JSON
+from sqlalchemy.orm import joinedload
 
 
 @OrgApp.html(model=OccurrenceCollection, template='occurrences.pt',
@@ -84,50 +83,57 @@ def ical_export_occurence(self, request):
     )
 
 
-@OrgApp.view(model=OccurrenceCollection, name='csv', permission=Public)
-def csv_export_occurences(self, request):
-    """ Returns all occurrence as csv. """
+@OrgApp.form(model=OccurrenceCollection, name='export', permission=Private,
+             form=ExportForm, template='export.pt')
+def view_export(self, request, form):
+    """ Export the occurrences in various formats. """
 
-    session = request.app.session()
+    layout = OccurrencesLayout(self, request)
+    layout.breadcrumbs.append(Link(_("Export"), '#'))
+    layout.editbar_links = None
 
-    # requires postgres >= 9.3
-    events = session.query(
-        Occurrence.title,
-        Occurrence.location,
-        func.json_extract_path_text(
-            func.cast(Event.content, JSON), 'organizer'
-        ),
-        func.array_to_string(func.akeys(Occurrence._tags), ','),
-        func.to_char(Occurrence.start, 'YYYY-MM-DD"T"HH24:MI:SS"Z"'),
-        func.to_char(Occurrence.end, 'YYYY-MM-DD"T"HH24:MI:SS"Z"'),
-        Occurrence.timezone,
-        func.json_extract_path_text(
-            func.cast(Event.content, JSON), 'description'
+    if form.submitted(request):
+        results = run_export(
+            request,
+            layout.export_formatter(form.format)
         )
-    )
-    events = events.outerjoin(Event)
-    events = events.filter(
-        Occurrence.start >= replace_timezone(as_datetime(date.today()), 'UTC')
-    )
-    events = events.order_by(Occurrence.start)
 
-    output = StringIO()
-    csv_writer = writer(output)
-    csv_writer.writerow((
-        'title',
-        'location',
-        'organizer',
-        'tags',
-        'start',
-        'end',
-        'timezone',
-        'description'
-    ))
-    writer(output).writerows(events)
-    output.seek(0)
+        return form.as_export_response(results, title=request.translate(_(
+            "Events"
+        )))
 
-    return Response(
-        output.read(),
-        content_type='text/csv',
-        content_disposition='inline; filename=events.csv'
+    return {
+        'layout': layout,
+        'title': _("Event Export"),
+        'form': form,
+        'explanation': _("Exports all future events.")
+    }
+
+
+def run_export(request, formatter):
+    occasions = OccurrenceCollection(request.app.session())
+
+    query = occasions.query(outdated=False)
+    query = query.options(joinedload(Occurrence.event))
+    query = query.order_by(Occurrence.start)
+
+    attributes = (
+        'title', 'location', 'event.organizer', 'tags',
+        'start', 'end', 'event.description'
     )
+
+    def get(obj, attr):
+
+        if attr == 'tags':
+            return [request.translate(_(tag)) for tag in obj.tags]
+
+        for a in attr.split('.'):
+            obj = value = getattr(obj, a)
+
+        return value
+
+    return [
+        OrderedDict(
+            (attr, formatter(get(occasion, attr))) for attr in attributes)
+        for occasion in query.all()
+    ]